Move code into a scheduler class: part 1
Move start-scheduler logic into a class (in engine.py). Call to
start-engine() will create a new engine class, and run a watchdog
thread on the cfg files associated with that engine.

In followup commits, will also start scheduler in the constructor,
and move more code into the class. The existing globals file can
be deleted at that point.

Change-Id: I3a547a538fecaabdb84c927df9439c30119bf74f