Do not create queues statically
Queues should be created programatically by audit scripts. Engine
should provide an exchange, and audit scripts can create queues
using that exchange.

Changed code in engine.py to create an exchange in the constructor.
Changed the base audit object to include a constructor that creates
queues.
Added an associated option to audit template to demonstrate this
feature.

Moved some code in start_audit to try/except blocks. Was a TODO

Change-Id: Ibe1e2283d6b182cd74b33c684a1477e5afe68f67