Make entropy suitable for pypi distribution, part 0
We need to separate out engine code and audit and repair scripts if
pypi distribution is reqd. This is handled in this commit.

Further commits will remove the two variables hardcoded currently,
and expose __main__ as a cmdline script on installation.

Use imp for module finding and loading, use full path to script in
audit and repair cfg files.

Move audit and repair scripts to examples. Make some changes to hardcoded
stuff to account for this

Update gitignore
Change-Id: I50831003c6f7272967dbeb5c558b76b0183c91be