add a wait next function to return next audit job
Add an engine timeout. This is the maximum time the engine will spin
without an audit job to run. Set in engine cfg.

Added a wait_next function. This will spin till an audit is available
or till timeout is reached. If there are jobs in the queue, return
all the jobs with the earliest execution time.

Change the way audit and react scripts are called. React scripts are
now started up from start_react_scripts at startup.

Remove some functions in engine that we don't use anymore.

Made some changes to audit.json. It doesn't make sense to have a name
in the config file, since we pass one in the commandline.

implements blueprint backend-abstraction
Change-Id: I14178a51b830b26856ea01dfb37cbfbdfa4ba703