Add kill logic
Add in logic to stop engines. To do this, add a field called enabled
in the engine cfg file that keeps track of files. Then add a new
parser to define the stop-engine CLI call, and a function to stop an
engine. This just does a psutil.terminate, which is equivalent to the
current steps to kill an engine.

Move cfg file creation to a function in utils

Change-Id: I1c365c2c438e6ed0a44413e1d09c69d3fab7ab7b