Move register-audit/repair code to backend
Abstract out cfg file operations in the register workflow. Changed
the get_driver code in engine to be static, so we can call it from
main too.

Added a new function in file_backend to return the config file given
the script type. Eg. return audit_cfg for audit.

Added a couple of string variables in base.py

The function get_cfg_file, when using a db, will actually return a
table. So rename it to something backend-agnostic

Change-Id: I20d6bd46caf56c750e4b1193a6f5d00ce4e930f6