Add callbacks for futures
Added callbacks for the known futures (react scripts, scheduler,
serializer). This will help us debug known long-lasting futures
easily. We need to add a new function for each kind of future,
because the futures module does not allow any args to be passed to
callbacks. This can be worked around using wrapper functions and the
such, but I dont expect to add any new kind of future so there wont
be any more duplication.

Change-Id: I579a41e92736568535ce3f64c5507ac5224d5a62